:truck: chore(api) move `CertificationScoringCompleted` event to `src`
