+++ api/lib/domain/events/handle-complementary-certifications-scoring.js
@@ -1,10 +1,10 @@
+import { CertificationScoringCompleted } from '../../../src/certification/evaluation/domain/events/CertificationScoringCompleted.js';
 import { ComplementaryCertificationCourseResult } from '../../../src/certification/shared/domain/models/ComplementaryCertificationCourseResult.js';
 import { AnswerCollectionForScoring } from '../../../src/shared/domain/models/AnswerCollectionForScoring.js';
 import { ComplementaryCertificationScoringWithComplementaryReferential } from '../../../src/shared/domain/models/ComplementaryCertificationScoringWithComplementaryReferential.js';
 import { ComplementaryCertificationScoringWithoutComplementaryReferential } from '../../../src/shared/domain/models/ComplementaryCertificationScoringWithoutComplementaryReferential.js';
 import { ReproducibilityRate } from '../../../src/shared/domain/models/ReproducibilityRate.js';
 import { CertificationRescoringCompleted } from './CertificationRescoringCompleted.js';
-import { CertificationScoringCompleted } from './CertificationScoringCompleted.js';
 import { checkEventTypes } from './check-event-types.js';
 
 const eventTypes = [CertificationScoringCompleted, CertificationRescoringCompleted];
